+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
+- Add `GitCommit` to the configuration
+- Add `DockerImageVersion` to the configuration
 
 ## 8.0.0 - 2019-11-05
 - [**BC**] Update ConsentsEvents and other dependencies